[librptexture] rp_image_backend: Reduce the struct size.
- rp_image::Format: Explicitly type as uint8_t.
- tr_idx: Reduce from int to int16_t.

On amd64, this reduces rp_image_backend from 32 bytes to 24 bytes.
On i386, this reduces rp_image_backend from 24 bytes to 20 bytes.

All unit tests are still passing.

+++ src/librptexture/img/rp_image_backend.hpp
@@ -56,16 +56,16 @@ class RP_LIBROMDATA_PUBLIC NOVTABLE rp_image_backend
 		int stride;
 		rp_image::Format format;
 
-		// Image data.
+		// Image data
 		virtual void *data(void) = 0;
 		virtual const void *data(void) const = 0;
 		virtual size_t data_len(void) const = 0;
 
-		// Image palette.
+		// Image palette
 		virtual uint32_t *palette(void) = 0;
 		virtual const uint32_t *palette(void) const = 0;
 		virtual unsigned int palette_len(void) const = 0;
-		int tr_idx;
+		int16_t tr_idx;
 
 	public:
 		// Subclasses can have other stuff here.
